# zz/swagger.yaml
openapi: 3.0.0
info:
  title: API Matchnhire Nodejs
  version: 1.0.0
  description: API de gestion des moyens et fiches de poste
servers:
  - url: http://localhost:3000/api
paths:
  /moyens:
    get:
      summary: Récupérer tous les moyens
      tags: [Moyens]
      responses:
        "200":
          description: Liste des moyens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Moyen"

    post:
      summary: Créer un moyen
      tags: [Moyens]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoyenInput"
      responses:
        "201":
          description: Moyen créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Moyen"
...
  /jobdescriptions:
    get:
      summary: Récupérer toutes les fiches de poste
      tags: [JobDescriptions]
      responses:
        "200":
          description: Liste des fiches de poste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobDescription"

    post:
      summary: Créer une fiche de poste
      tags: [JobDescriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobDescriptionInput"
      responses:
        "201":
          description: Fiche créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescription"
...
components:
  schemas:
    Moyen:
      type: object
      properties:
        id:
          type: integer
        moyen:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MoyenInput:
      type: object
      required:
        - moyen
      properties:
        moyen:
          type: string
    JobDescription:
      type: object
      properties:
        id:
          type: integer
        emploi:
          type: string
        filiere_activite:
          type: string
        famille:
          type: string
        superieur_n1:
          type: integer
        superieur_n2:
          type: integer
        finalite:
          type: string
        niveau_exigence:
          type: string
        niveau_exp:
          type: string
        status:
          type: string
        version:
          type: string
        missions:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        savoirs:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        savoirFaires:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        savoirEtres:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        moyens:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        airesProximites:
          type: array
          items:
            $ref: "#/components/schemas/AireProximite"
    JobDescriptionInput:
      type: object
      required: [emploi, filiere_activite, famille]
      properties:
        emploi:
          type: string
        filiere_activite:
          type: string
        famille:
          type: string
        superieur_n1:
          type: integer
        superieur_n2:
          type: integer
        finalite:
          type: string
        niveau_exigence:
          type: string
        niveau_exp:
          type: string
        status:
          type: string
        version:
          type: string
        missions:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        savoirs:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        savoirFaires:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        savoirEtres:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        moyens:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithId"
        airesProximites:
          type: array
          items:
            $ref: "#/components/schemas/AireProximite"
    ItemWithId:
      type: object
      properties:
        id:
          type: integer
        mission:
          type: string
        connaissance:
          type: string
        competence:
          type: string
        comportement:
          type: string
        niveau:
          type: string
        moyen:
          type: string
    AireProximite:
      type: object
      properties:
        id:
          type: integer
        poste:
          type: string
        nombre:
          type: integer